refactor: removing project attachment pages and merging into a single project revision page
